feat(chart): rename BarChart
BREAKING CHANGE: this bar chart is way too specific, so rename to OperationsHistoryBarChart to more clearly convey its intent.
