+++ src/components/charts/OperationsHistoryBarChart/Rect.jsx
